+++ internal/provider/resource_sdwan_service_dhcp_server_feature.go
@@ -64,7 +64,7 @@ func (r *ServiceDHCPServerProfileParcelResource) Metadata(ctx context.Context, r
 func (r *ServiceDHCPServerProfileParcelResource) Schema(ctx context.Context, req resource.SchemaRequest, resp *resource.SchemaResponse) {
 	resp.Schema = schema.Schema{
 		// This description is used by the documentation generator and the language server.
-		MarkdownDescription: helpers.NewAttributeDescription("This resource can manage a Service DHCP Server Feature.").AddMinimumVersionDescription("20.12.0").String,
+		MarkdownDescription: helpers.NewAttributeDescription("This resource can manage a Service DHCP Server Feature.").AddMinimumVersionDescription("20.15.0").String,
 
 		Attributes: map[string]schema.Attribute{
 			"id": schema.StringAttribute{
@@ -93,6 +93,9 @@ func (r *ServiceDHCPServerProfileParcelResource) Schema(ctx context.Context, req
 			"network_address": schema.StringAttribute{
 				MarkdownDescription: helpers.NewAttributeDescription("Network Address").String,
 				Optional:            true,
+				Validators: []validator.String{
+					stringvalidator.RegexMatches(regexp.MustCompile(`^((25[0-5]|2[0-4]\d|1\d\d|[1-9]\d|\d)\.){3}(25[0-5]|2[0-4]\d|1\d\d|[1-9]\d|\d)$`), ""),
+				},
 			},
 			"network_address_variable": schema.StringAttribute{
 				MarkdownDescription: helpers.NewAttributeDescription("Variable name").String,