+++ docs/guides/changelog.md
@@ -11,6 +11,7 @@ description: |-
 
 - BREAKING CHANGE: Bump the minimum supported version of `sdwan_application_priority_qos_policy` to `20.15.0`
 - BREAKING CHANGE: Bump the minimum supported version of `sdwan_dns_security_policy` to `20.15.0`
+- BREAKING CHANGE: Bump the minimum supported version of `sdwan_service_dhcp_server_feature` to `20.15.0`
 
 ## 0.10.0